Improve agreement styles and formatting
Now the page uses whatever the system font is. We also make the page
more responsive and use a font size that's a better fit for the content.
This change also modifies the raw text to be more grammatically correct
(if that's a thing) and read better. Finally, the styles were tweaked
for print layout so that everything would fit neatly onto one page.